creation of super_admin at initialization of db + user_cli complete

+++ epic/cli/user_cli.py
@@ -1,4 +1,4 @@
-from epic.models.models import User, Role, create_tables
+from epic.models.models import User, Role
 import typer
 from peewee import DoesNotExist
 from epic.cli.auth_cli import authenticated_command
@@ -10,26 +10,26 @@ import inspect
 app = typer.Typer()
 
 method_allowed = {
-    "user_cli.create_user": ["admin"],
-    "user_cli.list_users": ["admin"],
-    "user_cli.delete_user": ["admin"],
-    "user_cli.update_user": ["admin"],
-    "user_cli.create_role": ["admin"],
-    "user_cli.list_roles": ["admin"],
-    "user_cli.delete_role": ["admin"],
-    "user_cli.update_role": ["admin"],
-    "event_cli.create_event": ["admin", "sales", "support"],
-    "event_cli.list_events": ["admin", "sales", "support"],
-    "event_cli.delete_event": ["admin", "sales", "support"],
-    "event_cli.update_event": ["admin", "sales", "support"],
-    "client_cli.create_client": ["admin", "sales"],
-    "client_cli.list_clients": ["admin", "sales"],
-    "client_cli.delete_client": ["admin", "sales"],
-    "client_cli.update_client": ["admin", "sales"],
-    "contract_cli.create_contract": ["admin", "sales"],
-    "contract_cli.list_contracts": ["admin", "sales"],
-    "contract_cli.delete_contract": ["admin", "sales"],
-    "contract_cli.update_contract": ["admin", "sales"],
+    "user_cli.create_user": ["admin", "super_admin"],
+    "user_cli.list_users": ["admin", "super_admin"],
+    "user_cli.delete_user": ["admin", "super_admin"],
+    "user_cli.update_user": ["admin", "super_admin"],
+    "user_cli.create_role": ["super_admin"],
+    "user_cli.list_roles": ["super_admin"],
+    "user_cli.delete_role": ["super_admin"],
+    "user_cli.update_role": ["super_admin"],
+    "event_cli.create_event": ["admin", "sales", "support", "super_admin"],
+    "event_cli.list_events": ["admin", "sales", "support", "super_admin"],
+    "event_cli.delete_event": ["admin", "sales", "support", "super_admin"],
+    "event_cli.update_event": ["admin", "sales", "support", "super_admin"],
+    "client_cli.create_client": ["admin", "sales", "super_admin"],
+    "client_cli.list_clients": ["admin", "sales", "super_admin"],
+    "client_cli.delete_client": ["admin", "sales", "super_admin"],
+    "client_cli.update_client": ["admin", "sales", "super_admin"],
+    "contract_cli.create_contract": ["admin", "sales", "super_admin"],
+    "contract_cli.list_contracts": ["admin", "sales", "super_admin"],
+    "contract_cli.delete_contract": ["admin", "sales", "super_admin"],
+    "contract_cli.update_contract": ["admin", "sales", "super_admin"],
 }
 
 
@@ -55,13 +55,12 @@ def create_user():
 
     Example:
         To create a new user with the name "John Doe", email "<EMAIL>", password "password", and role "admin", you can run the following command:
-        $ python app.py user create
+        $ python -m epic user create
         Enter name: John Doe
         Enter email: <EMAIL>
         Enter password: password
         Enter role name: admin
         User John Doe created successfully."""
-    # Check permission
     function_name = inspect.currentframe().f_code.co_name
     if user_info["role"] in method_allowed[filename + "." + function_name]:
         name = typer.prompt("Enter name:")
@@ -79,8 +78,24 @@ def create_user():
 
 
 @app.command("list")
-@authenticated_command
 def list_users():
+    """Get a list of all users in the system
+
+    Args:
+    None
+
+    Returns:
+    A list of all users in the system
+
+    Raises:
+    None
+
+    Example:
+    To get a list of all users in the system, you can run the following command:
+    $ python -m epic user list
+    User ID: 1, Name: John Doe, Email: <EMAIL>, Role: admin
+    User ID: 2, Name: Jane Doe, Email: <EMAIL>, Role: user
+    ..."""
     users = User.select()
     for user in users:
         typer.echo(
@@ -91,7 +106,23 @@ def list_users():
 @app.command("delete")
 @authenticated_command
 def delete_user():
-    # Check permission
+    """Deletes a user from the system.
+
+    Args:
+        user_id (int): The ID of the user to delete.
+
+    Returns:
+        None
+
+    Raises:
+        DoesNotExist: If the user with the specified ID does not exist.
+
+    Example:
+        To delete a user with the ID of 1, you can run the following command:
+        $ python -m epic user delete
+        1
+        User John Doe deleted successfully."""
+
     function_name = inspect.currentframe().f_code.co_name
     if user_info["role"] in method_allowed[filename + "." + function_name]:
         user_id = typer.prompt("Enter user ID:")
@@ -108,7 +139,33 @@ def delete_user():
 @app.command("update")
 @authenticated_command
 def update_user():
-    # Check permission
+    """Updates an existing user.
+
+    This function prompts the user to enter the ID of the user to update, their new name, email, password, and role name, and then updates the user with the provided information.
+
+    Args:
+        user_id (int): The ID of the user to update.
+
+    Returns:
+        None
+
+    Raises:
+        DoesNotExist: If the user with the specified ID does not exist.
+
+    Example:
+        To update a user with the ID of 1 with the name "John Doe", email "<EMAIL>", password "password", and role "admin", you can run the following command:
+        $ python -m epic user update
+        1
+        Enter new name or press 'Enter':
+        John Doe
+        Enter new email or press 'Enter':
+        <EMAIL>
+        Enter new password or press 'Enter':
+        password
+        Enter new role name or press 'Enter':
+        admin
+        User John Doe updated successfully."""
+
     function_name = inspect.currentframe().f_code.co_name
     if user_info["role"] in method_allowed[filename + "." + function_name]:
         user_id = typer.prompt("Enter user ID to update:")
@@ -147,14 +204,54 @@ def update_user():
 
 @app.command("create-role")
 @authenticated_command
-def create_role(name: str):
-    role = Role.create(name=name)
-    typer.echo(f"Role {role.name} created successfully.")
+def create_role():
+    """Creates a new role.
+
+    This function prompts the user to enter the name of the role to create, and then creates a new role with the provided name.
+
+    Args:
+        name (str): The name of the role to create.
+
+    Returns:
+        None
+
+    Raises:
+        DoesNotExist: If a role with the same name already exists.
+
+    Example:
+        To create a new role with the name "manager", you can run the following command:
+        $ python -m epic user create-role
+        Enter role name: manager
+        Role manager created successfully."""
+    function_name = inspect.currentframe().f_code.co_name
+    if user_info["role"] in method_allowed[filename + "." + function_name]:
+        name = typer.prompt("Enter role name:")
+        role = Role.create(name=name)
+        typer.echo(f"Role {role.name} created successfully.")
+    else:
+        print("User not allowed")
 
 
 @app.command("list-roles")
 @authenticated_command
 def list_roles():
+    """Get a list of all roles in the system.
+
+    Args:
+        None
+
+    Returns:
+        A list of all roles in the system
+
+    Raises:
+        None
+
+    Example:
+        To get a list of all roles in the system, you can run the following command:
+        $ python -m epic user list-roles
+        Role ID: 1, Name: admin
+        Role ID: 2, Name: support
+        ..."""
     roles = Role.select()
     for role in roles:
         typer.echo(f"Role ID: {role.id}, Name: {role.name}")
@@ -162,27 +259,35 @@ def list_roles():
 
 @app.command("delete-role")
 @authenticated_command
-def delete_role(role_id: int):
-    try:
-        role = Role.get(Role.id == role_id)
-        role.delete_instance()
-        typer.echo(f"Role {role.name} deleted successfully.")
-    except DoesNotExist:
-        typer.echo(f"Role with ID {role_id} does not exist.")
+def delete_role():
+    """Deletes a role from the system.
 
+    Args:
+        role_id (int): The ID of the role to delete.
 
-@app.command("update-role")
-@authenticated_command
-def update_role(role_id: int, name: str):
-    try:
-        role = Role.get(Role.id == role_id)
-        role.name = name
-        role.save()
-        typer.echo(f"Role {role.name} updated successfully.")
-    except DoesNotExist:
-        typer.echo(f"Role with ID {role_id} does not exist.")
+    Returns:
+        None
+
+    Raises:
+        DoesNotExist: If the role with the specified ID does not exist.
+
+    Example:
+        To delete a role with the ID of 1, you can run the following command:
+        $ python -m epic user delete-role 1
+        Role with ID 1 deleted successfully."""
+
+    function_name = inspect.currentframe().f_code.co_name
+    if user_info["role"] in method_allowed[filename + "." + function_name]:
+        role_id = typer.prompt("Enter role ID:")
+        try:
+            role = Role.get(Role.id == role_id)
+            role.delete_instance()
+            typer.echo(f"Role {role.name} deleted successfully.")
+        except DoesNotExist:
+            typer.echo(f"Role with ID {role_id} does not exist.")
+    else:
+        print("User not allowed")
 
 
 if __name__ == "__main__":
-    create_tables()
-    app()
+    pass
